red-team(w2.3): P0 add external_connections.account_id (RPC crashed) + test type fixes
pgTAP surfaced two real bugs once the plan parsed:

P0 (migration): plaid_ingest_transactions reads & writes
external_connections.account_id, but the migration never added that column — every
ingest raised 42703 "column account_id does not exist". Prod would have crashed on
the first bank sync. Added the column (uuid → ledger_accounts) alongside sync_cursor.

Test: the ledger-balance assertions compared sum()::numeric to 0::bigint
(is(numeric,bigint) has no operator). Cast the journal_lines sum to ::bigint.

+++ supabase/tests/w2_3_plaid_ingest_test.sql
@@ -54,7 +54,7 @@ select is((select count(distinct journal_entry_id)::int from bank_transactions w
 -- the ledger balances (every posted entry Dr==Cr — sum of all lines nets to 0).
 select is(
   (select coalesce(sum(case when side='D' then amount_minor else -amount_minor end),0)
-   from journal_lines where org_id='00000000-0000-0000-0000-0000000e0b01'),
+   from journal_lines where org_id='00000000-0000-0000-0000-0000000e0b01')::bigint,
   0::bigint, 'ledger balances after add (Dr==Cr across all lines)');
 
 -- ── 3. W2.3-REPLAY: re-ingesting the same page is a NO-OP (webhook replay) ────
@@ -93,7 +93,7 @@ select is(
 -- ledger still balances after the reversal + repost.
 select is(
   (select coalesce(sum(case when side='D' then amount_minor else -amount_minor end),0)
-   from journal_lines where org_id='00000000-0000-0000-0000-0000000e0b01'),
+   from journal_lines where org_id='00000000-0000-0000-0000-0000000e0b01')::bigint,
   0::bigint, 'ledger still balances after modify (reverse + repost)');
 
 -- ── 5. W2.3-REMOVED: removing t2 reverses its entry, marks the row removed ────
